requirements.txt
django>=5.2.0,<5.2.7
djangorestframework
